Restore original contact form
This patch restores the original contact form so that we can get all of
the fields on a single page.  The multi step process with typeform is
really annoying and a poor experience.

A simple form sets the stage for something like formspree.io to receive
the post data and forward the form as an email.