Add comments to posts using disqus
This is the basic disqus comment box.  It will load on ever page render,
which may really slow the site down.  There's also the UX question of
comments.  I'm thinking I'm going to take the middle ground on UX and
load time and do something like this [1] to only load comments on the
click of a button.

[1]: http://www.paulund.co.uk/ajax-disqus-comment-system