Move about, services, contact to single page with separate blog
This patch does two things.  First, it adds author information and integrates
metadata to Google profiles in an effort to get author images showing up in
search results.  These are proven ways to get more traffic.

Second, the layout of the site moves to a single page approach with sections for
services offered, about, and contact us.

The blog itself moves to /blog/ and will itself be a multi-page site with
navigation back to the main landing page of the single page site.